api/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
websockets==11.0.3
httpx==0.25.0
asyncpg==0.29.0
aiofiles==23.2.1
